moved pages to correct folder
